// src/types/fish.ts
export type Fish = {
    id: number;
    name: string;
    ph: string;
    kh: string;
    gh: string;
    temperature: string;
    img: string;
    mediumSize?: number;
    maxSize?: number;
    longevity?: number;
    diet?: string;
    sociability?: string;
    territoriality?: string;
    wayOfLife?: string;
    wayOfBreeding?: string;
    stream?: string;
    minimumVolume?: number;
};